Separate Omid-related config from HBase config
A new class OmidClientConfiguration has been created in the
transaction-client module in order to encompass all the
configuration related to Omid client (was included before in TSOClient)
This class is passed in the builder of the TSOClient

Similarly a new class -HBaseOmidClientConfiguration- has been created
to encompass the previous config plus the specific config options of Omid for
HBase (e.g. the commit table name). This class is passed in the builder
of the HBaseTransactionManager

This simplifies the configuration of the majority of tests and the user
examples. It also simplifies the wiring in the HBaseTransactionManager
builder and the creation of TSOClient instances.

NOTE: No functional changes have been added in this commit.