This commit is related to jira [OMID-72].
This commit adds support for fences over a table. These fences guarantee that every transaction started before the fence and writes to the table the fence was created for will be aborted.
These fences are needed for creating a secondary index in Apache Phoenix. The scenario was discussed in TEPHRA-157 and PHOENIX-2478. Augmenting Omid with this kind of transactions was also discussed in OMID-56.